refactor: move physical planning context to own module
Move PhysicalPlanningContext and its scalar subquery state types out of execution_props.rs and update imports and migration guidance.

AI Disclosure: This code was written in part by an AI agent.
