+++ datafusion/core/src/execution/context/mod.rs
@@ -2383,7 +2383,7 @@ mod tests {
     use arrow_schema::FieldRef;
     use datafusion_common::DataFusionError;
     use datafusion_common::datatype::DataTypeExt;
-    use datafusion_expr::execution_props::PhysicalPlanningContext;
+    use datafusion_expr::physical_planning_context::PhysicalPlanningContext;
     use std::error::Error;
     use std::path::PathBuf;
