+++ datafusion/core/src/test_util/parquet.rs
@@ -28,7 +28,8 @@ use crate::datasource::listing::{ListingTableUrl, PartitionedFile};
 use crate::datasource::object_store::ObjectStoreUrl;
 use crate::datasource::physical_plan::ParquetSource;
 use crate::error::Result;
-use crate::logical_expr::execution_props::{ExecutionProps, PhysicalPlanningContext};
+use crate::logical_expr::execution_props::ExecutionProps;
+use crate::logical_expr::physical_planning_context::PhysicalPlanningContext;
 use crate::logical_expr::simplify::SimplifyContext;
 use crate::optimizer::simplify_expressions::ExprSimplifier;
 use crate::physical_expr::create_physical_expr;
