+++ datafusion/functions-nested/src/lambda_utils.rs
@@ -138,9 +138,10 @@ pub(crate) mod test_utils {
     use datafusion_common::{DFSchema, Result};
     use datafusion_expr::{
         Expr, HigherOrderUDF, col,
-        execution_props::{ExecutionProps, PhysicalPlanningContext},
+        execution_props::ExecutionProps,
         expr::{HigherOrderFunction, LambdaVariable},
         lambda,
+        physical_planning_context::PhysicalPlanningContext,
     };
     use datafusion_physical_expr::create_physical_expr;
