+++ docs/source/library-user-guide/upgrading/55.0.0.md
@@ -660,7 +660,7 @@ uncorrelated scalar-subquery state to functions that create physical
 `Arc<dyn PhysicalExpr>` values from logical `Expr` values.
 
 That state is now carried by a dedicated
-`datafusion_expr::execution_props::PhysicalPlanningContext` passed explicitly
+`datafusion_expr::physical_planning_context::PhysicalPlanningContext` passed explicitly
 through functions and planner traits. Unlike `ExecutionProps`, which applies
 throughout the planning of an entire query, this context is scoped to the
 logical plan subtree currently being converted. This removes the need for the
@@ -708,7 +708,7 @@ When creating a physical expression outside of physical planning, pass an empty
 context:
 
 ```rust,ignore
-use datafusion_expr::execution_props::PhysicalPlanningContext;
+use datafusion_expr::physical_planning_context::PhysicalPlanningContext;
 use datafusion_physical_expr::create_physical_expr;
 
 // Before
